FIXME: Test failure:

CURRENT_TEST: innodb.innodb-alter
mysqltest: At line 252: query 'ALTER TABLE t1o ADD COLUMN FTS_DOC_ID BIGINT UNSIGNED NOT NULL,
ALGORITHM=INPLACE' failed with wrong errno 1815: 'Internal error: InnoDB: ADD COLUMN...INSTANT', instead of 1846...

+++ include/my_bit.h
@@ -17,8 +17,6 @@
 #ifndef MY_BIT_INCLUDED
 #define MY_BIT_INCLUDED
 
-#include <my_global.h>
-
 /*
   Some useful bit functions
 */
@@ -119,9 +117,9 @@ static inline uint32 my_reverse_bits(uint32 key)
   a number with the n lowest bits set
   an overflow-safe version of  (1 << n) - 1
 */
-static inline uint32 my_set_bits(int n)
+static inline uint64 my_set_bits(int n)
 {
-  return (((1UL << (n - 1)) - 1) << 1) | 1;
+  return (((1ULL << (n - 1)) - 1) << 1) | 1;
 }
 
 C_MODE_END
